Matrix Addition excercise
Matrix Addition:
Write a program that creates three matrix variables of dimensions 2x3, two of which
are to be initialized in such a way that they are identical to A and B; the third 
will be used to store the sum of A and B. Using double for-loop to iterate over the
matrix elements, compute the sum of each component and store results in the third 
matrix. Finally, output the matrix sum. Program's output should look like the following:

A =
-5 2 8
1 0 0 

B = 
1 0 2 
0 3 -6 

A + B =
-4 2 10
1 3 -6
Press any key to continue...