Linear Search excercise
Linear Search:
Hardcode the following integer array: {7, 3, 32, 2, 55, 34, 6, 13, 29, 22, 11, 9, 1, 5, 42, 39, 8}
into your program. Display this array to the user. Then ask the user to input an integer to search for.
Program should then search the array for the integer the user entered and ouput its array position
(array index). Output should look similar to the following:

List = 7, 3, 32, 2, 55, 34, 6, 13, 29, 22, 11, 9, 1, 5, 42, 39, 8
Enter an integer in the list to search for: 55
Item found at index [4]
Press any key to continue...